+++ src/day14/chemicalTypes.kt
@@ -1,7 +1,7 @@
 package day14
 
 typealias Chemical = String
-typealias ReactionResult = Pair<Chemical, Int>
+typealias ReactionResult = Pair<Chemical, Long>
 typealias ReactionComponent = ReactionResult
 typealias ReactionComponents = List<ReactionComponent>
 typealias ReactionResultLookup = MutableMap<Chemical, ReactionResult>
